♻️ (approvals): Extract filter helpers to filter.ts to satisfy react-refresh
